django/requirements.txt
Django==4.1.11
djongo==1.3.6
pymongo==3.12.3
pytz
djangorestframework